fixed bug with _staticcall that didn't forward to logic contract correctly We need to do a hack to get around delegatecall being a non-view function, for some reason the fallback function can delegatecall view functions and return data.
So we'll staticcall ourselves with a function selector of
0xffffffff so we always hit the fallback function and then
do a delegatecall in there and return the value to _staticcall

This, somehow, keeps _staticcall a view function but also makes sure
data cant be changed when doing delegatecall?

This is not very gas efficent, however there's no other way to allow
_staticcall to perform a delegatecall to the logic contract and still
remain a view function.